+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2024-03-13
+### Added
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `Material UI` library.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputString` and `InputAreaString` components.
+
 ## [1.14.0] - 2024-03-11
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `/meal-proposal` page for proposals of the best meals.